+++ app/src/main/java/com/example/playlistmaker/media/ui/MediaViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.singleActivity
+package com.example.playlistmaker.media.ui
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
