+++ app/src/main/res/layout/fragment_single_media.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:orientation="vertical"
     android:layout_height="match_parent"
-    tools:context=".singleActivity.FragmentSingleMedia">
+    tools:context=".media.ui.FragmentSingleMedia">
 
     <com.google.android.material.tabs.TabLayout
         android:id="@+id/tab_layout"
