+++ app/src/main/res/layout/fragment_tab_favourites.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center_horizontal"
     android:orientation="vertical"
-    tools:context=".singleActivity.FragmentTabFavourites">
+    tools:context=".media.ui.FragmentTabFavourites">
 
     <ImageView
         android:layout_width="120dp"
