+++ app/src/main/res/layout/fragment_tab_playlists.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center_horizontal"
     android:orientation="vertical"
-    tools:context=".singleActivity.FragmentTabPlaylists">
+    tools:context=".media.ui.FragmentTabPlaylists">
 
     <Button
         style="@style/reloadButton"
